Add support for --auto-decompress-gz option
When this option is given any file that ends with ".gz" is automatically
decompressed before added to the (compressed) archive.

This can give a substantial improvement in compression rates. Not only
because brotli is a better compressor than gzip; if the different
decompressed .gz files are similar brotli can also see the redundancy
across the different files.